# packages/api-spec/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/milestones/{id}:
    put:
      tags:
      - milestone-controller
      operationId: updateMilestone
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMilestoneResponse"
    delete:
      tags:
      - milestone-controller
      operationId: deleteMilestone
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    patch:
      tags:
      - milestone-controller
      operationId: updateMilestoneStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneStatusUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMilestoneResponse"
  /api/labels/{id}:
    get:
      tags:
      - label-controller
      operationId: getLabel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLabelDetailResponse"
    put:
      tags:
      - label-controller
      operationId: updateLabel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLabelDetailResponse"
    delete:
      tags:
      - label-controller
      operationId: deleteLabel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/users/signup:
    post:
      tags:
      - user-controller
      operationId: postSignupRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/users/signin:
    post:
      tags:
      - user-controller
      operationId: postSignInRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAccessTokenResponse"
  /api/users/refresh:
    post:
      tags:
      - user-controller
      operationId: refreshAccessToken
      parameters:
      - name: refreshToken
        in: cookie
        required: false
        schema:
          type: string
          default: none
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAccessTokenResponse"
  /api/milestones:
    get:
      tags:
      - milestone-controller
      operationId: getMilestoneList
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - OPEN
          - CLOSED
          default: OPEN
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMilestoneListResponse"
    post:
      tags:
      - milestone-controller
      operationId: postNewMilestone
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMilestoneResponse"
  /api/labels:
    get:
      tags:
      - label-controller
      operationId: getLabels
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLabelListResponse"
    post:
      tags:
      - label-controller
      operationId: createLabel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLabelDetailResponse"
  /api/issues:
    get:
      tags:
      - issue-controller
      operationId: mainPage
      parameters:
      - name: condition
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/IssueSearchCondition"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseIssueSearchResponse"
    post:
      tags:
      - issue-controller
      operationId: createIssue
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseIssueDetailResponse"
  /api/issues/{id}/comments:
    get:
      tags:
      - comment-controller
      operationId: getCommentsForIssue
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCommentListResponse"
    post:
      tags:
      - comment-controller
      operationId: postCommentForIssue
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCommentResponse"
  /api/issues/{id}:
    get:
      tags:
      - issue-controller
      operationId: issueDetail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseIssueDetailResponse"
    patch:
      tags:
      - issue-controller
      operationId: updateStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueStatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/issues/status:
    patch:
      tags:
      - issue-controller
      operationId: bulkUpdateStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkIssueRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/users/me:
    get:
      tags:
      - user-controller
      operationId: getMyInfo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserInfoResponse"
  /api/comments/{id}:
    delete:
      tags:
      - comment-controller
      operationId: deleteCommentByCommentId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
components:
  schemas:
    MilestoneRequest:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 0
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date
    ApiResponseMilestoneResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/MilestoneResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    ErrorDto:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    MilestoneResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
        openIssueCount:
          type: integer
          format: int32
        closedIssueCount:
          type: integer
          format: int32
    LabelRequest:
      required:
      - backgroundColor
      - name
      - textColor
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        backgroundColor:
          pattern: "^#[A-Fa-f0-9]{6}$"
          type: string
        textColor:
          type: string
          enum:
          - DARK
          - LIGHT
    ApiResponseLabelDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/LabelDetailResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    LabelDetailResponse:
      type: object
      properties:
        labelId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        backgroundColor:
          type: string
        textColor:
          type: string
          enum:
          - DARK
          - LIGHT
    SignupRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 32
          minLength: 4
          type: string
        password:
          maxLength: 64
          minLength: 8
          pattern: "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!]).*$"
          type: string
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          $ref: "#/components/schemas/ErrorDto"
    LoginRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 32
          minLength: 4
          type: string
        password:
          maxLength: 64
          minLength: 8
          pattern: "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!]).*$"
          type: string
    AccessTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
    ApiResponseAccessTokenResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AccessTokenResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    IssueRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        labelIds:
          type: array
          items:
            type: integer
            format: int64
        milestoneId:
          type: integer
          format: int64
    ApiResponseIssueDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/IssueDetailResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    IssueDetailResponse:
      type: object
      properties:
        issueNumber:
          type: integer
          format: int64
        title:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
        createdAt:
          type: string
          format: date-time
        labels:
          type: array
          items:
            $ref: "#/components/schemas/LabelSummaryResponse"
        milestone:
          $ref: "#/components/schemas/MilestoneSummaryResponse"
    LabelSummaryResponse:
      type: object
      properties:
        labelId:
          type: integer
          format: int64
        name:
          type: string
        backgroundColor:
          type: string
        textColor:
          type: string
          enum:
          - DARK
          - LIGHT
    MilestoneSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        openIssueCount:
          type: integer
          format: int32
        closedIssueCount:
          type: integer
          format: int32
    CommentRequest:
      type: object
      properties:
        content:
          type: string
    ApiResponseCommentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/CommentResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time
    MilestoneStatusUpdateRequest:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
    UpdateIssueStatusRequest:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
    BulkIssueRequest:
      required:
      - issueIds
      - status
      type: object
      properties:
        issueIds:
          type: array
          items:
            type: integer
            format: int64
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
    ApiResponseUserInfoResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/UserInfoResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    UserInfoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        profileImageUrl:
          type: string
    ApiResponseMilestoneListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/MilestoneListResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    MilestoneListResponse:
      type: object
      properties:
        milestoneCount:
          type: integer
          format: int32
        openMilestoneCount:
          type: integer
          format: int32
        closedMilestoneCount:
          type: integer
          format: int32
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/MilestoneResponse"
    ApiResponseLabelListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/LabelListResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    LabelListResponse:
      type: object
      properties:
        labels:
          type: array
          items:
            $ref: "#/components/schemas/LabelDetailResponse"
    IssueSearchCondition:
      type: object
      properties:
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
    ApiResponseIssueSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/IssueSearchResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    IssueSearchResponse:
      type: object
      properties:
        openIssueCount:
          type: integer
          format: int64
        closedIssueCount:
          type: integer
          format: int64
        issues:
          type: array
          items:
            $ref: "#/components/schemas/IssueSummaryResponse"
    IssueSummaryResponse:
      type: object
      properties:
        issueNumber:
          type: integer
          format: int64
        title:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
        createdAt:
          type: string
          format: date-time
        labels:
          type: array
          items:
            $ref: "#/components/schemas/LabelSummaryResponse"
        milestone:
          $ref: "#/components/schemas/MilestoneReferenceResponse"
    MilestoneReferenceResponse:
      type: object
      properties:
        milestoneId:
          type: integer
          format: int64
        title:
          type: string
    ApiResponseCommentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/CommentListResponse"
        error:
          $ref: "#/components/schemas/ErrorDto"
    CommentListResponse:
      type: object
      properties:
        issueNumber:
          type: integer
          format: int64
        comment_count:
          type: integer
          format: int32
        comments:
          type: array
          items:
            $ref: "#/components/schemas/CommentResponse"